add the jwt parts

+++ part3/hbnb/app/api/v1/reviews.py
@@ -1,8 +1,8 @@
 #!/usr/bin/python3
-
-
 from flask_restx import Namespace, Resource, fields
-from app.services import facade
+from hbnb.app.services import facade
+from flask_jwt_extended import jwt_required, get_jwt_identity
+
 
 api = Namespace('reviews', description='Review operations')
 
@@ -17,13 +17,32 @@ review_model = api.model('Review', {
 
 @api.route('/')
 class ReviewList(Resource):
+    @jwt_required()
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
     def post(self):
         """Register a new review"""
+        current_user = get_jwt_identity()  # task 3: Identify currently logged-in user
         review_data = api.payload
+
+        # Check if user is owner of place
+        place = facade.get_place(review_data['place_id'])
+        if not place:
+            return 'Error Not Found', 404
+
+        # Owner is not allowed to review their own place
+        if place.owner_id == current_user.get("id"):
+            return 'You cannot review your own place.', 400
+
+        # Check if user already left a review
+        review = facade.get_reviews_by_place(review_data['place_id'])
+        for review in review:
+            if review.user_id == current_user.get("id"):
+                return 'You have already reviewed this place..', 400
+
         try:
+            review_data['user_id'] = current_user.get("id")
             new_review = facade.create_review(review_data)
 
             # Serialize the review
@@ -67,6 +86,7 @@ class ReviewList(Resource):
 
 @api.route('/<review_id>')
 class ReviewResource(Resource):
+    @jwt_required()
     @api.response(200, 'Review details retrieved successfully')
     @api.response(404, 'Review not found')
     def get(self, review_id):
@@ -87,13 +107,21 @@ class ReviewResource(Resource):
         else:
             return {'error': 'Review does not exist'}, 404
 
+    @jwt_required()
     @api.expect(review_model)
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
     def put(self, review_id):
         """Update a review's information"""
+        current_user = get_jwt_identity()
         review_data = api.payload
+
+        review = facade.get_review(review_id)
+
+        if review.user_id != current_user.get("id"):
+            return {'error': 'Unauthorized action.'}, 403
+
         try:
             updated_review = facade.update_review(review_id, review_data)
             if updated_review:
@@ -113,10 +141,17 @@ class ReviewResource(Resource):
         except Exception as e:
             return {'error': str(e)}, 400
 
+    @jwt_required()
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
     def delete(self, review_id):
         """Delete a review"""
+        current_user = get_jwt_identity()
+        review = facade.get_review(review_id)
+
+        if review.user_id != current_user.get("id"):
+            return {'error': 'Unauthorized action.'}, 403
+
         try:
             delete_review = facade.delete_review(review_id)
             if delete_review:
